Swal alerts and some minor corrections

+++ src/components/Photo.jsx
@@ -9,6 +9,8 @@ import { useDispatch } from "react-redux";
 import DeleteOutlineIcon from "@mui/icons-material/DeleteOutline";
 import { saveAs } from "file-saver";
 import NestedModal from "./NestedModal";
+import Swal from 'sweetalert2'
+
 
 export const Photo = (props) => {
   const dispatch = useDispatch();
@@ -27,7 +29,24 @@ export const Photo = (props) => {
     }
   };
   const removefavorite = () => {
-    dispatch(removePhoto(props.item));
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(removePhoto(props.item));
+        Swal.fire(
+          'Deleted!',
+          'Your photo has been deleted.',
+          'success'
+        )
+      }
+    })
   };
  
   return (
